Fix bug where field was not recoverable due to noise
If PAL_DO_VSYNC was disabled and the noise was high, the field was being incorrectly determined since it was still scanning the noisy signal to find out which field it is. This patch makes it so if PAL_DO_VSYNC is disabled, the field is determined before noise is added to the signal.